Optimisation/preparations for filename encryption
We're now loading all the encrypted filenames in a key/value store upon
start for the commands we need it for. Future work will include reusing
matching folder names, as every time you encrypt a string it's
different. This is good if we don't want to disclose we have the same
names in different folders, but bad for reusing the existing folder
structure.

Also preliminary support for inserting a password with an encrypted
path.